CE complete for 1001-Deploy network workloads.md
Hi @ckezema, I've completed the edit pass for 1001-Deploy network workloads.md

A few comments:
Line 84, 152
"After the installation succeeded message displays, select **Close**."

Is there a specific message? Or is “installation succeeded” the actual message?

Line 361
"3. At the SConfig management screen, enter `15` and then select Enter to exit to the command line (PowerShell)."

Per our style, it should be: Server Configuration tool (Sconfig) for first mention. However, if you’re referring to it as it displays in the UI, then it’s ok to leave it.

Line 391
   > This will return information about three rules, including one for HTTP and one for HTTPS. Verify that both rules are enabled and allow inbound traffic.

You list only 2 rules. I suggest changing this to “This will return information about two rules, including…”. Or if it really is three rules, remove the word “about”. You could replace it with “approximately” if it’s variable.