CE complete for 0201-Manage disks and volumes in Windows Server.md
Hi @ckezema,

I've completed my edit pass of 0201-Manage disks and volumes in Windows Server.md, and it is ready for your review. In addition:
Line 173. Can we change to "Notice the NTFS volume on **Disk 5** has a size of approximately **10** gigabytes (GB)."
Line 213: Can we change to "...for use in Microsoft Hyper-V running in Windows Server."