CE complete for Mod 3 LAK - Implementing enterprise storage solutions.md
Hi @ckezema,

I've completed my review for this Mod 3 LAK, "Implementing enterprise storage solutions," and it's ready for your review. Comments are:
Line 48: Can we change this to "...storage for Microsoft SQL Server databases?" If you agree to change, please search for the remaining "SQL database (or databases), and change to "SQL Server database/databases.
Line 271. Instead of "at the message," can we say what the box or message are? Something more descriptive would be nice.
Line 401: Should Windows be Windows Server?