CE complete for Mod 5 LAK, Install and configure Hyper-V
Hi @ckezema,

I've completed the edits for the mod 5 LAK "Install and configure Hyper-V," and it's ready for your review. A few comments in here though:

Line 65: Can we change "...you should have successfully verified..." to "you successfully verified," or "you will have verified?"
Line 90: Can we give them a name for the drop-down menu? Someone who cannot see otherwise might not be able to find it unless it is intuitive.
Line 97: Where do they select Local Server from, in Server Manager?
Line 112: Can we change "...has been added. A private network..." to "...has been added, and that a private network...?" Otherwise, should the sentence beginning w/ "A private network" be changed to a Note?
Line 172: Can we change **New Virtual Machine** **Wizard** to **New Virtual Machine Wizard**? Same comment for lines 174, 200, and 248.
Lines 336-340: By "area," do we mean "pane?"
Line 350: Can we change "because it was a production checkpoint" to a Note? For example, "**LON-GUEST2** status is off, because it was a production checkpoint.
Line 433: Consider changing "The VM might restart several times." to a Note.
Line 439: Should this be, "At the command prompt, enter **15**, and then select ..."
Line 441: Should **Get-WindowsFeature Hyper** be **Get-WindowsFeature Hyper-V**?
Line 451: Can we change "you should have" to either "you will have" or "you configured a ..?"